feat: add flag migration logic
This allows us to migrate flag files to settings.json, as well as have dual read support for our init.d scripts.

Note that dump_emmc has some special logic to clear it's own flag, since it needs this and can't use x1plusd since it runs so early in init.

We also need a copy jq on the internal fs for our boot script to check the flag, so move it to prebuilt and symlink elsewhere.